+++ tests/unit/tasks/test_callable_task.py
